backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
PyMuPDF
python-multipart